+++ tests/QueryBuilderTest.php
@@ -218,6 +218,32 @@ class QueryBuilderTest extends TestCase
         $this->assertEmpty($binding);
     }
 
+    public function testMakeCount_withDistinct()
+    {
+        $queryBuilder = new FakeQueryBuilder($this->getConnection());
+        $queryBuilder = $queryBuilder
+            ->column('f.bar')
+            ->from('foo', 'f')
+            ->orderBy('bar', 'DESC')
+            ->limit(2);
+        $binding = [];
+
+        $select = $queryBuilder->makeCount();
+        $this->assertEquals(
+            'SELECT COUNT(*) AS `count` FROM ( SELECT 1 FROM foo AS f ) AS countable',
+            $select->getStatement($binding)
+        );
+        $this->assertEmpty($binding);
+
+        $queryBuilder->distinct();
+        $select = $queryBuilder->makeCount();
+        $this->assertEquals(
+            'SELECT COUNT(*) AS `count` FROM ( SELECT DISTINCT f.bar FROM foo AS f ) AS countable',
+            $select->getStatement($binding)
+        );
+        $this->assertEmpty($binding);
+    }
+
     public function testMakeCount_withHaving()
     {
         $queryBuilder = new FakeQueryBuilder($this->getConnection());